Bugfix
CRITICAL bug fix
RemoveCardFromInventoryAction to only delete cards belonging to the currently authenticated user, preventing removal of cards from other users' inventories.

+++ app/Actions/Card/RemoveCardFromInventoryAction.php
@@ -4,6 +4,7 @@ namespace App\Actions\Card;
 
 use App\Actions\Action;
 use App\Models\UserCard;
+use Illuminate\Support\Facades\Auth;
 use App\Actions\Contracts\ActionInterface;
 
 class RemoveCardFromInventoryAction extends Action implements ActionInterface
@@ -11,7 +12,8 @@ class RemoveCardFromInventoryAction extends Action implements ActionInterface
     public function perform(...$args)
     {
         $card_id = $args[0];
+        $user_id = Auth::id();
 
-        UserCard::where('card_id', '=', $card_id)->delete();
+        UserCard::where('card_id', '=', $card_id)->where('user_id', '=', $user_id)->delete();
     }
 }
